backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
numpy
pydicom
Pillow
